refactor(mespapiers): Rename some Home components
no code change
Home/Content => Home/HomeLayout
Home/Result => Home/ResultWithoutSearch
Home/HomeSearchResult => Home/ResultForSearch
Home/PapersList => Home/Content
